+++ src/chess/board.py
@@ -1,17 +1,13 @@
-"""Chess"""
+__name__ = "Chess board module"
+__all__ = ["Board"]
 
-
-from enum import Enum
 from itertools import product
-
-
-class Color(Enum):
-    WHITE = 1
-    BLACK = 2
+from .utils import *
+from .pieces import *
 
 
 class Board:
-    """Main chess board"""
+    """Main chess board class"""
 
     def __init__(self):
         self.check: Color | None = None
@@ -132,10 +128,10 @@ class Board:
                 if isinstance(piece, King):
                     if self.can_castle(row, col, row1, col1):
                         piece.set_moved()
-                        self.color = opponent(self.color)
+                        self.color = self.color.opponent()
                         return self.castle(row, col, row1, col1)
                 return False
-        elif dest.get_color() == opponent(piece.get_color()):
+        elif dest.get_color() == piece.get_color().opponent():
             if not piece.can_attack(self, row, col, row1, col1):
                 return False
         else:
@@ -143,7 +139,7 @@ class Board:
         piece.set_moved()
         self.field[row][col] = None
         self.field[row1][col1] = piece
-        self.color = opponent(self.color)
+        self.color = self.color.opponent()
         self.check_check()
         return True
 
@@ -217,7 +213,7 @@ class Board:
         self.field[row][col] = None
         self.field[row1][col1] = new_piece
         self.check_check()
-        self.color = opponent(self.color)
+        self.color = self.color.opponent()
         return True
 
     def check_check(self) -> None:
@@ -229,8 +225,8 @@ class Board:
                 if king_piece is None:
                     continue
                 if isinstance(king_piece, King):
-                    if self.is_under_attack(i, j, opponent(king_piece.get_color())):
-                        self.check = opponent(king_piece.get_color())
+                    if self.is_under_attack(i, j, king_piece.get_color().opponent()):
+                        self.check = king_piece.get_color().opponent()
                         self.mate_check(i, j, king_piece)
 
     def get_check(self) -> Color | None:
@@ -246,7 +242,7 @@ class Board:
             ):  # or king_piece.can_attack(self, row, col, row + i, col + j):
                 self.mate = None
                 return
-        self.mate = opponent(king_piece.get_color())
+        self.mate = king_piece.get_color().opponent()
 
     def get_mate(self) -> Color | None:
         """Returns current mate state"""
@@ -345,329 +341,3 @@ class Board:
         self.field[row][col1].set_moved()
 
         return True
-
-
-class Piece:
-    """Abstract Piece"""
-
-    def __init__(self, color) -> None:
-        self.color = color
-        self._moved = False
-
-    def get_color(self) -> Color:
-        """Returns piece color"""
-        return self.color
-
-    def char(self) -> str:
-        """Returns piece representation"""
-        ...
-
-    def __str__(self) -> str:
-        col = "w" if self.color == Color.WHITE else "b"
-        return col + self.char()
-
-    def can_move(self, board: Board, row: int, col: int, row1: int, col1: int) -> bool:
-        """Check for abstract move possibility"""
-        if not correct_coords(row1, col1):
-            return False
-        if row == row1 and col == col1:
-            return False
-        piece = board.get_piece(row1, col1)
-        if piece is not None:
-            if piece.get_color() == self.color:
-                return False
-        return True
-
-    def can_attack(
-        self, board: Board, row: int, col: int, row1: int, col1: int, team_check=True
-    ) -> bool:
-        """Check for abstract attack possibility"""
-        if not correct_coords(row1, col1):
-            return False
-        if row == row1 and col == col1:
-            return False
-        piece = board.get_piece(row1, col1)
-        if piece is not None:
-            if piece.get_color() == self.color and team_check:
-                return False
-        return True
-
-    def set_moved(self):
-        self._moved = True
-
-    def moved(self):
-        return self._moved
-
-
-class Rook(Piece):
-    """Rook"""
-
-    def get_color(self) -> int:
-        return self.color
-
-    def char(self) -> str:
-        return "R"
-
-    def can_move(self, board: Board, row: int, col: int, row1: int, col1: int) -> bool:
-        if not super().can_move(board, row, col, row1, col1):
-            return False
-        if row != row1 and col != col1:
-            return False
-
-        step = 1 if (row1 >= row) else -1
-        for _r in range(row + step, row1, step):
-            if board.get_piece(_r, col) is not None:
-                return False
-
-        step = 1 if (col1 >= col) else -1
-        for _c in range(col + step, col1, step):
-            if board.get_piece(row, _c) is not None:
-                return False
-        return True
-
-    def can_attack(
-        self, board: Board, row: int, col: int, row1: int, col1: int, team_check=True
-    ) -> bool:
-        if not super().can_attack(board, row, col, row1, col1):
-            return False
-        return self.can_move(board, row, col, row1, col1)
-
-
-class Pawn(Piece):
-    """Pawn"""
-
-    def get_color(self) -> int:
-        return self.color
-
-    def char(self) -> str:
-        return "P"
-
-    def can_move(self, board: Board, row: int, col: int, row1: int, col1: int) -> bool:
-        if not super().can_move(board, row, col, row1, col1):
-            return False
-        if col != col1:
-            return False
-
-        if board.get_piece(row1, col1) is not None:
-            return False
-
-        if self.color == Color.WHITE:
-            direction = -1
-            start_row = 6
-        else:
-            direction = 1
-            start_row = 1
-
-        if row + direction == row1:
-            return True
-
-        if (
-            row == start_row
-            and row + 2 * direction == row1
-            and board.field[row + direction][col] is None
-        ):
-            return True
-        return False
-
-    def can_attack(
-        self, board: Board, row: int, col: int, row1: int, col1: int, team_check=True
-    ) -> bool:
-        if not super().can_attack(board, row, col, row1, col1):
-            return False
-        direction = -1 if (self.color == Color.WHITE) else 1
-        return row + direction == row1 and (col + 1 == col1 or col - 1 == col1)
-
-
-class Knight(Piece):
-    """Knight"""
-
-    def char(self) -> str:
-        return "N"
-
-    def can_move(self, board: Board, row: int, col: int, row1: int, col1: int) -> bool:
-        if not super().can_move(board, row, col, row1, col1):
-            return False
-        delta_row = abs(row - row1)
-        delta_col = abs(col - col1)
-
-        if sorted([delta_row, delta_col]) == [1, 2]:
-            return True
-        return False
-
-    def can_attack(
-        self, board: Board, row: int, col: int, row1: int, col1: int, team_check=True
-    ) -> bool:
-        if not super().can_attack(board, row, col, row1, col1):
-            return False
-        return self.can_move(board, row, col, row1, col1)
-
-
-class Bishop(Piece):
-    """Bishop"""
-
-    def char(self) -> str:
-        return "B"
-
-    def can_move(self, board: Board, row: int, col: int, row1: int, col1: int) -> bool:
-        if not super().can_move(board, row, col, row1, col1):
-            return False
-        delta_row = row - row1
-        delta_col = col - col1
-
-        if abs(delta_row) != abs(delta_col):
-            return False
-
-        step_x = 1 if col < col1 else -1
-        step_y = 1 if row < row1 else -1
-
-        for i in range(1, abs(delta_row)):
-            _row = i * step_y + row
-            _col = i * step_x + col
-            if not correct_coords(_row, _col):
-                continue
-
-            if board.get_piece(_row, _col) is not None:
-                return False
-
-        return True
-
-    def can_attack(
-        self, board: Board, row: int, col: int, row1: int, col1: int, team_check=True
-    ) -> bool:
-        if not super().can_attack(board, row, col, row1, col1):
-            return False
-        return self.can_move(board, row, col, row1, col1)
-
-
-class Queen(Piece):
-    """Queen"""
-
-    def char(self) -> str:
-        return "Q"
-
-    def can_move(self, board: Board, row: int, col: int, row1: int, col1: int) -> bool:
-        if not super().can_move(board, row, col, row1, col1):
-            return False
-
-        if col == col1:
-            step = 1 if (row1 >= row) else -1
-            for _row in range(row + step, row1, step):
-                if board.get_piece(_row, col) is not None:
-                    return False
-            return True
-
-        if row == row1:
-            step = 1 if (col1 >= col) else -1
-            for _col in range(col + step, col1, step):
-                if board.get_piece(row, _col) is not None:
-                    return False
-            return True
-
-        delta_row = row - row1
-        delta_col = col - col1
-
-        if abs(delta_row) == abs(delta_col):
-            step_x = 1 if col < col1 else -1
-            step_y = 1 if row < row1 else -1
-
-            for i in range(1, abs(delta_row)):
-                _row = i * step_y + row
-                _col = i * step_x + col
-                if not correct_coords(_row, _col):
-                    continue
-
-                if board.get_piece(_row, _col) is not None:
-                    return False
-            return True
-        return False
-
-    def can_attack(
-        self, board: Board, row: int, col: int, row1: int, col1: int, team_check=True
-    ) -> bool:
-        if not super().can_attack(board, row, col, row1, col1):
-            return False
-        return self.can_move(board, row, col, row1, col1)
-
-
-class King(Piece):
-    """King"""
-
-    def char(self) -> str:
-        return "K"
-
-    def can_move(self, board: Board, row: int, col: int, row1: int, col1: int) -> bool:
-        if not super().can_move(board, row, col, row1, col1):
-            return False
-        move_x = abs(col - col1)
-        move_y = abs(row - row1)
-        if move_x not in [0, 1]:
-            return False
-        if move_y not in [0, 1]:
-            return False
-        if board.is_under_attack(row1, col1, opponent(self.color)):
-            return False
-        return True
-
-    def can_attack(
-        self, board: Board, row: int, col: int, row1: int, col1: int, team_check=True
-    ) -> bool:
-        if not super().can_attack(board, row, col, row1, col1):
-            return False
-        return self.can_move(board, row, col, row1, col1)
-
-
-def opponent(color: Color):
-    """Returns opposite color"""
-    if color == Color.WHITE:
-        return Color.BLACK
-    return Color.WHITE
-
-
-def correct_coords(row, col):
-    """Check if coords is inside board bounds"""
-    return 0 <= row < 8 and 0 <= col < 8
-
-
-def print_board(board):
-    """Output board to console"""
-    print("     +----+----+----+----+----+----+----+----+")
-    for row in range(7, -1, -1):
-        print(" ", row, end="  ")
-        for col in range(8):
-            print("|", board.cell(row, col), end=" ")
-        print("|")
-        print("     +----+----+----+----+----+----+----+----+")
-    print(end="        ")
-    for col in range(8):
-        print(col, end="    ")
-    print()
-
-
-def main():
-    """main"""
-    board = Board()
-
-    while True:
-        print_board(board)
-
-        print(
-            f"Commands:\n\texit\t\t\t\t-- exit\n\tmove <row> <col> <row1> <col1>\t\t\t\t-- move from (row, col) to (row1, col1)"
-        )
-
-        if board.current_player_color() == Color.WHITE:
-            print("Turn of white:")
-        else:
-            print("Turn of black:")
-        command = input()
-        if command == "exit":
-            break
-        _, row, col, row1, col1 = command.split()
-        row, col, row1, col1 = int(row), int(col), int(row1), int(col1)
-        if board.move_piece(row, col, row1, col1):
-            print("Turn succeeded")
-        else:
-            print("Wrong coords! Try again!")
-
-
-if __name__ == "__main__":
-    main()
