+++ src/chess/pieces/__init__.py
@@ -1,0 +1,10 @@
+"""Pieces module"""
+
+
+from .piece import Piece
+from .king import King
+from .pawn import Pawn
+from .bishop import Bishop
+from .rook import Rook
+from .queen import Queen
+from .knight import Knight
