+++ src/ui/__init__.py
@@ -1,0 +1,2 @@
+from .pawn_promotion_dialog import PawnPromotionDialog
+from .chess_window import ChessWindow
